NOTE: THE BUILD MAY BE BROKEN.
I have added "refCon" parameters to many of the chained calls, so that we can allow the caller to attach a refCon.

I will make them default to nil, eventually, but I want to deliberately break the build, for now, so I get all the places it needs to be supplied.